finished presentation layer, added animations spacing etc

+++ app/src/main/java/com/example/wachanga/data/datasource/NoteDao.kt
@@ -15,7 +15,7 @@ interface NoteDao {
     fun getNotes(): Flowable<List<Note>>
 
     @Query("SELECT * from note WHERE id = :id")
-    fun getNoteById(id: Int): Note?
+    fun getNoteById(id: Long): Note?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     fun insertNote(note: Note): Long
